requirements.txt
# ============================================================================
# Touchless Media Control - Dependencies
# NVIDIA Jetson Nano Developer Kit (4GB)
# Python 3.6 | MediaPipe 0.8.5 | CUDA 10.2 | OpenCV 4.1.1
# ============================================================================

# --- Core Computer Vision & ML ---
# MediaPipe 0.8.5 ARM64 wheel - download from https://github.com/PINTO0309/mediapipe-bin
# mediapipe==0.8.5
numpy>=1.16.0,<1.20.0

# --- Configuration ---
PyYAML>=5.1,<6.0

# --- System Control ---
# xdotool is installed via apt: sudo apt install xdotool

# --- Performance Monitoring ---
psutil>=5.6.0

# ============================================================================
# JETSON NANO INSTALLATION NOTES:
#
# 1. OpenCV 4.1.1 is pre-installed with JetPack 4.x:
#      python3 -c "import cv2; print(cv2.__version__)"
#
# 2. MediaPipe 0.8.5 ARM64 wheel for Jetson Nano:
#      wget https://github.com/PINTO0309/mediapipe-bin/releases/download/v0.8.5/
#           mediapipe-0.8.5_cuda102-cp36-cp36m-linux_aarch64.whl
#      pip3 install mediapipe-0.8.5_cuda102-cp36-cp36m-linux_aarch64.whl
#
# 3. CUDA 10.2 is pre-installed with JetPack 4.x
#
# 4. VLC keyboard control:
#      sudo apt install xdotool vlc
#
# 5. Jetson performance monitoring:
#      sudo pip3 install jetson-stats
#      sudo jtop  (to monitor GPU/CPU/thermal)
# ============================================================================
